Release kyc separate

+++ lib/sila/models/kyb/check_kyb_response/verification_history.dart
@@ -8,7 +8,7 @@ class VerificationHistory extends Equatable {
   final int updatedAt;
   final List<Null> reasons;
   final List<String> tags;
-  final int score;
+  final double score;
   final String parentVerification;
 
   VerificationHistory(
